CI-frontend/requirements.txt
streamlit
Pillow
streamlit-option-menu
requests
load_dotenv
pandas